+++ src/relstorage/adapters/postgresql/mover.py
@@ -13,12 +13,13 @@
 ##############################################################################
 """IObjectMover implementation.
 """
-from __future__ import absolute_import
+from __future__ import absolute_import, print_function
 
 import functools
 import os
 
 from zope.interface import implementer
+from ZODB.POSException import Unsupported
 
 from ..._compat import xrange
 from .._util import query_property
@@ -26,23 +27,55 @@ from ..interfaces import IObjectMover
 from ..mover import AbstractObjectMover
 from ..mover import metricmethod_sampled
 
+# Important: pg8000 1.10 - 1.13, at least, can't handle prepared
+# statements that take parameters but it doesn't need to because it
+# prepares every statement anyway. So you must have a backup that you use
+# for that driver.
+# https://github.com/mfenniak/pg8000/issues/132
 
-def to_prepared_queries(name, queries, datatypes=''):
-    # Only handles one param
-    return [
-        'PREPARE ' + name + ' ' + datatypes + ' AS ' + x.replace('%s', '$1')
-        for x in queries
-    ]
+
+def to_prepared_queries(name, queries, datatypes=()):
+    # Give correct datatypes for the queries, wherever possible.
+    # The number of parameters should be the same or more than the
+    # number of datatypes.
+    # datatypes is a sequence of strings.
+
+    # Maybe instead of having the adapter have to know about all the
+    # statements that need prepared, we could keep a registry?
+    if datatypes:
+        assert isinstance(datatypes, (list, tuple))
+        datatypes = ', '.join(datatypes)
+        datatypes = ' (%s)' % (datatypes,)
+    else:
+        datatypes = ''
+
+    result = []
+    for q in queries:
+        if not isinstance(q, str):
+            # Unsupported marker
+            result.append(q)
+            continue
+
+        q = q.strip()
+        param_count = q.count('%s')
+        rep_count = 0
+        while rep_count < param_count:
+            rep_count += 1
+            q = q.replace('%s', '$' + str(rep_count), 1)
+        stmt = 'PREPARE {name}{datatypes} AS {query}'.format(
+            name=name, datatypes=datatypes, query=q
+        )
+        result.append(stmt)
+    return result
+
 
 @implementer(IObjectMover)
 class PostgreSQLObjectMover(AbstractObjectMover):
 
-    __need_version_check = True
-
     _prepare_load_current_queries = to_prepared_queries(
         'load_current',
         AbstractObjectMover._load_current_queries,
-        '(BIGINT)')
+        ['BIGINT'])
 
     _prepare_load_current_query = query_property('_prepare_load_current')
 
@@ -56,86 +89,112 @@ class PostgreSQLObjectMover(AbstractObjectMover):
 
     _detect_conflict_query = 'EXECUTE detect_conflicts'
 
-    on_load_opened_statement_names = ('_prepare_load_current_query',)
-    on_store_opened_statement_names = on_load_opened_statement_names + (
-        '_prepare_detect_conflict_query',)
-
-    # Sadly we can't PREPARE this statement; apparently it holds a
-    # lock on OBJECT_STATE that interferes with taking the commit lock.
-    _move_from_temp_object_state_95_query = """
+    _move_from_temp_hf_insert_query_raw = """
         INSERT INTO object_state (zoid, tid, state_size, state)
         SELECT zoid, %s, COALESCE(LENGTH(state), 0), state
         FROM temp_store
-        ON CONFLICT (zoid) DO UPDATE SET state_size = COALESCE(LENGTH(excluded.state), 0),
-                              tid = %s,
-                              STATE = excluded.state
+        ON CONFLICT (zoid)
+        DO UPDATE
+        SET state_size = COALESCE(LENGTH(excluded.state), 0),
+            tid = excluded.tid,
+            state = excluded.state
     """
 
-    def _move_from_temp_object_state_95(self, cursor, tid):
-        stmt = self._move_from_temp_object_state_95_query
-        cursor.execute(stmt, (tid, tid))
+    _move_from_temp_hf_insert_raw_queries = (
+        Unsupported("States accumulate in history-preserving mode"),
+        _move_from_temp_hf_insert_query_raw,
+    )
+
+    _prepare_move_from_temp_hf_insert_queries = to_prepared_queries(
+        'move_from_temp',
+        _move_from_temp_hf_insert_raw_queries,
+        ('BIGINT',)
+    )
+
+    _prepare_move_from_temp_hf_insert_query = query_property(
+        '_prepare_move_from_temp_hf_insert')
+
+    _move_from_temp_hf_insert_queries = (
+        Unsupported("States accumulate in history-preserving mode"),
+        'EXECUTE move_from_temp(%s)'
+    )
+
+    _move_from_temp_hf_insert_query = query_property('_move_from_temp_hf_insert')
+
+    on_load_opened_statement_names = ('_prepare_load_current_query',)
+    on_store_opened_statement_names = on_load_opened_statement_names + (
+        '_prepare_detect_conflict_query',
+        '_prepare_move_from_temp_hf_insert_query',
+    )
+
 
     @metricmethod_sampled
     def on_store_opened(self, cursor, restart=False):
         """Create the temporary tables for storing objects"""
         # note that the md5 column is not used if self.keep_history == False.
-        stmts = [
+        # Ideally we wouldn't execute any of these on a restart, but
+        # I've seen an issue with temp_stare apparently going missing on pg8000
+        ddl_stmts = [
             """
-            CREATE TEMPORARY TABLE temp_store (
-                zoid        BIGINT NOT NULL,
+            CREATE TEMPORARY TABLE IF NOT EXISTS temp_store (
+                zoid        BIGINT NOT NULL PRIMARY KEY,
                 prev_tid    BIGINT NOT NULL,
                 md5         CHAR(32),
                 state       BYTEA
-            ) ON COMMIT DROP;
+            ) ON COMMIT DELETE ROWS;
             """,
             """
-            CREATE UNIQUE INDEX temp_store_zoid ON temp_store (zoid);
-            """,
-            """
-            CREATE TEMPORARY TABLE temp_blob_chunk (
+            CREATE TEMPORARY TABLE IF NOT EXISTS temp_blob_chunk (
                 zoid        BIGINT NOT NULL,
                 chunk_num   BIGINT NOT NULL,
-                chunk       OID
-            ) ON COMMIT DROP;
-            """,
-            """
-            CREATE UNIQUE INDEX temp_blob_chunk_key
-            ON temp_blob_chunk (zoid, chunk_num);
-            """,
-            """
-            -- This trigger removes blobs that get replaced before being
-            -- moved to blob_chunk.  Note that it is never called when
-            -- the temp_blob_chunk table is being dropped or truncated.
-            CREATE TRIGGER temp_blob_chunk_delete
-                BEFORE DELETE ON temp_blob_chunk
-                FOR EACH ROW
-                EXECUTE PROCEDURE temp_blob_chunk_delete_trigger();
+                chunk       OID,
+                PRIMARY KEY (zoid, chunk_num)
+            ) ON COMMIT DELETE ROWS;
             """,
         ]
-        for stmt in stmts:
-            cursor.execute(stmt)
+        if not restart:
+            ddl_stmts += [
+                """
+                -- This trigger removes blobs that get replaced before being
+                -- moved to blob_chunk.  Note that it is never called when
+                -- the temp_blob_chunk table is being dropped or truncated.
+                CREATE TRIGGER temp_blob_chunk_delete
+                    BEFORE DELETE ON temp_blob_chunk
+                    FOR EACH ROW
+                    EXECUTE PROCEDURE temp_blob_chunk_delete_trigger();
+                """,
+            ]
+            # For some reason, preparing the INSERT statement also wants
+            # to acquire a lock. If we're committing is another
+            # transaction, this can block indefinitely (if that other transaction
+            # happens to be in this same thread!)
+            # checkIterationIntraTransaction (PostgreSQLHistoryPreservingRelStorageTests)
+            # easily triggers this. Fortunately, I don't think this
+            # is a common case, and we can workaround the test failure by
+            # only prepping this in the store connection.
+            # TODO: Is there a more general solution?
+            cursor.execute('SET lock_timeout = 100')
 
-        if self.__need_version_check:
-            self.__need_version_check = False
-            supports_conflict = self.version_detector.get_version(cursor) >= (9, 5)
-            if supports_conflict:
-                self._move_from_temp_object_state = self._move_from_temp_object_state_95
-                self.store_temp = self._store_temp_95
+        for stmt in ddl_stmts:
+            cursor.execute(stmt)
 
         AbstractObjectMover.on_store_opened(self, cursor, restart)
 
-    def _store_temp_95(self, _cursor, batcher, oid, prev_tid, data):
+    def _move_from_temp_object_state(self, cursor, tid):
+        # Override the history-free version of moving from
+        # temp_store to do it in one step.
+        stmt = self._move_from_temp_hf_insert_query
+        cursor.execute(stmt, (tid,))
 
+
+    @metricmethod_sampled
+    def store_temp(self, _cursor, batcher, oid, prev_tid, data):
         suffix = """
         ON CONFLICT (zoid) DO UPDATE SET state = excluded.state,
                               prev_tid = excluded.prev_tid,
                               md5 = excluded.md5
         """
         self._generic_store_temp(batcher, oid, prev_tid, data, suffix=suffix)
-
-    @metricmethod_sampled
-    def store_temp(self, cursor, batcher, oid, prev_tid, data): # pylint:disable=method-hidden
-        self._generic_store_temp(batcher, oid, prev_tid, data)
 
     @metricmethod_sampled
     def restore(self, cursor, batcher, oid, tid, data):
@@ -272,11 +331,13 @@ class PostgreSQLObjectMover(AbstractObjectMover):
 
 
 class PG8000ObjectMover(PostgreSQLObjectMover):
-    # pg8000 1.10 can't handle prepared statements that take parameters
-    # but it doesn't need to because it prepares every statement
-    # anyway. https://github.com/mfenniak/pg8000/issues/132
-
+    # Delete the statements that need paramaters.
     on_load_opened_statement_names = ()
     on_store_opened_statement_names = ('_prepare_detect_conflict_query',)
 
     _load_current_query = AbstractObjectMover._load_current_query
+
+    _move_from_temp_hf_insert_queries = (
+        Unsupported("States accumulate in history-preserving mode"),
+        PostgreSQLObjectMover._move_from_temp_hf_insert_query_raw
+    )