+++ CHANGES.rst
@@ -5,8 +5,16 @@
 3.0a2 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Drop support for PostgreSQL versions earlier than 9.6. See
+  :issue:`220`.
 
+- Make MySQL and PostgreSQL use a prepared statement to get
+  transaction IDs. PostgreSQL also uses a prepared statement to set
+  them. This may be slightly faster. See :issue:`246`.
+
+- Make PostgreSQL use a prepared statement to move objects to their
+  final destination during commit (history free only). See
+  :issue:`246`.
 
 3.0a1 (2019-06-12)
 ==================